Chapter20/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pillow>=8.0.0
onnx>=1.12.0
onnxruntime>=1.12.0
azureml-core>=1.48.0
azure-ai-ml>=1.8.0
azure-identity>=1.12.0
inference-schema[numpy-support]>=1.5.0
six>=1.16.0 